Address PR feedback: move bank files to real_environment/, drop dead config
- Move bank_state.py, bank_tools.py, bank_tools_unsafe.py into
  suites/minibank/real_environment/ to clarify these simulate a
  customer's existing backend (not something suite authors write)
- Add README explaining the suite structure and real_environment purpose
- Drop ground_truth, ground_truth_output, placeholder_args from suite.yaml
  (removed from parser in #52)
